refactor: consolidate duplicate markdown/frontmatter utils into src/utils/markdown.js
